requirements.txt
asgiref==3.8.1
boto3==1.34.127
botocore==1.34.127
certifi==2024.2.2
charset-normalizer==3.3.2
diff-match-patch==20230430
dj-database-url==2.2.0
Django==5.0.6
django-background-tasks==1.2.5
django-cleanup==8.1.0
django-compat==1.0.15
django-import-export==4.0.7
django-jalali-date==1.1.2
django-storages==1.14.3
idna==3.7
jalali_core==1.0.0
jdatetime==5.0.0
jmespath==1.0.1
pillow==10.3.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.2
s3transfer==0.10.1
six==1.16.0
sqlparse==0.5.0
tablib==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
